mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.51.0",
    "asgiref>=3.8.1",
    "flask>=3.1.1",
    "mcp>=1.9.0",
    "nba-api>=1.9.0",
    "python-dotenv>=1.1.0",
]
